// src/store/arena/types.ts
export enum Types {
  CONTEST_REGISTRATION_SUCCEED = 'CONTEST_REGISTRATION_SUCCEED',
  CONTEST_REGISTRATION_FAILED = 'CONTEST_REGISTRATION_FAILED',
  RESET_CONTEST_REGISTRATION = 'RESET_CONTEST_REGISTRATION',
  GET_CANDIDATE_SUCCEED = 'GET_CANDIDATE_SUCCEED',
  SAVE_DATA_ROUND_INFO = 'SAVE_DATA_ROUND_INFO',
  IS_OPEN_EDIT_USER = 'IS_OPEN_EDIT_USER',
  TAB_OPTION = 'TAB_OPTION',
  IS_FORM_CODE = 'IS_FORM_CODE',
  GET_ANSWER = 'GET_ANSWER',
  SET_ANSWER = 'SET_ANSWER',
  SET_MUTI_ANSWER = 'SET_MUTI_ANSWER'
}